reworked grid, cell, and algoritms core
- added abstract base class for algorithms
- added cell data property
- modified some algorithms to iclude basic logging
- modified all algorithms to use the base class
- made sure all the examples still work
- probably broke a bunch of mypy types on the way but I'll fix that soon

+++ base/cell.py
@@ -1,4 +1,5 @@
-from typing import cast, Dict, List, Optional    # noqa: F401
+from random import choice
+from typing import Dict, List, Optional, Tuple, cast  # noqa: F401
 
 from base.distances import Distances
 
@@ -6,35 +7,41 @@ from base.distances import Distances
 class Cell:
 
     @property
-    def row(self) -> int:
-        return self._row
+    def links(self):
+        return self._links
 
-    @property
-    def col(self) -> int:
-        return self._col
-
-    @property
-    def links(self) -> List["Cell"]:
+    @links.getter
+    def links(self):
         return list(self._links.keys())
 
     @property
-    def neighbors(self) -> List["Cell"]:
-        neighbors_list = []         # type: List[Cell]
-        if self.north:
-            neighbors_list.append(self.north)
-        if self.south:
-            neighbors_list.append(self.south)
-        if self.east:
-            neighbors_list.append(self.east)
-        if self.west:
-            neighbors_list.append(self.west)
-        return neighbors_list
+    def neighbours(self) -> List['Cell']:
+        ''' List of neighbours '''
+        neighbours = []         # type: List[Cell]
+        if self.north: neighbours.append(self.north)
+        if self.south: neighbours.append(self.south)
+        if self.east:  neighbours.append(self.east)
+        if self.west:  neighbours.append(self.west)
+        return neighbours
+    
+    def randomNeighbour(self) -> Optional['Cell']:
+        ''' Return a random neighbour '''
+        if self.nn > 0: return choice(self.neighbours)
+    
+    @property
+    def nl(self) -> int:
+        ''' Number of links '''
+        return len(self.links)
+    
+    @property
+    def nn(self) -> int:
+        ''' Number of neighbours '''
+        return len(self.neighbours)
 
     @property
     def distances(self) -> Distances:
         distances = Distances(self)
         frontier = [self]
-
         while len(frontier) > 0:
             new_frontier = []
             for cell in frontier:
@@ -43,7 +50,6 @@ class Cell:
                         distances[linked_cell] = cast(int, distances[cell]) + 1
                         new_frontier.append(linked_cell)
             frontier = new_frontier
-
         return distances
 
     def __init__(self, row: int, col: int) -> None:
@@ -52,39 +58,73 @@ class Cell:
         if col is None or col < 0:
             raise ValueError("Column must be a positive integer")
 
-        self._row = row         # type: int
-        self._col = col   # type: int
+        self.row = row          # type: int
+        self.col = col          # type: int
         self._links = {}        # type: Dict[Cell, bool]
+        self._data = {}         # type: Dict
         self.north = None       # type: Optional[Cell]
         self.south = None       # type: Optional[Cell]
         self.east = None        # type: Optional[Cell]
         self.west = None        # type: Optional[Cell]
 
-    def link(self, cell: "Cell", bidirectional: bool = True) -> "Cell":
+    def link(self, cell: 'Cell', bidi: bool = True) -> None:
+        ''' Link yourself to another cell '''
         self._links[cell] = True
-        if bidirectional:
-            cell.link(self, False)
+        if bidi: cell.link(self, False)
+    
+    def __iadd__(self, cell: 'Cell') -> None:
+        ''' Overload for the += operator with the link method '''
+        self.link(cell)
         return self
 
-    def unlink(self, cell: "Cell", bidirectional: bool = True) -> "Cell":
-        del self._links[cell]
-        if bidirectional:
-            cell.unlink(self, False)
+    def unlink(self, cell: 'Cell', bidi: bool = True) -> None:
+        ''' Unlink yourself from another cell '''
+        if self.linked(cell): del self._links[cell]
+        if bidi: cell.unlink(self, False)
+
+    def __isub__(self, cell: 'Cell') -> None:
+        ''' Overload for the -= operator with the unlink method '''
+        self.unlink(cell)
         return self
 
-    def linked_to(self, cell: "Cell") -> bool:
-        return cell in self._links
+    def linked(self, cell: 'Cell') -> bool:
+        ''' Check if this cell is linked to another '''
+        return cell != None and cell in self.links
+
+    def __and__(self, other: 'Cell') -> bool:
+        ''' Overload for the & operator with the linked? method '''
+        return self.linked(other)
+
+    @property
+    def data(self):
+        ''' Accesses the data dictionary '''
+        return self._data
+
+    # def setData(self, key, value) -> None:
+    #     ''' Sets cell data '''
+    #     self._data[key] = value
+
+    def hasData(self, key) -> bool:
+        ''' Checks wheter cell contains key '''
+        return key in self.data.keys()
+
+    # def getData(self, key):
+    #     ''' Gets data value if present '''
+    #     if self.hasData(key): return self._data[key]
+
+    def __hash__(self) -> int:
+        ''' Unique hash of the cell '''
+        return hash((self.col, self.row, id(self)))
 
     # The following methods actually lie because don't take into account neighbors/linked-cells, but for now is enough
 
-    def __hash__(self) -> int:
-        return hash((self.col, self.row))
-
     def __repr__(self) -> str:
-        return "({},{})".format(self.col, self.row)
+        ''' print representation '''
+        ID = str(hex(id(self)))
+        return 'Cell at ({},{}) with memory address of {}'.format(self.row, self.col, ID)
 
-    def __eq__(self, other_cell: "Cell") -> bool:       # type: ignore
-        return self.row == other_cell.row and self.col == other_cell.col
+    def __eq__(self, other: 'Cell') -> bool:       # type: ignore
+        return hash(self) == hash(other)
 
-    def __ne__(self, other_cell: "Cell") -> bool:       # type: ignore
-        return not self == other_cell
+    def __ne__(self, other: 'Cell') -> bool:       # type: ignore
+        return not self == other
